requirements.txt
openai==1.3.5
numpy==1.24.3
scikit-learn==1.2.2
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.0.331
pydantic==2.4.2
gradio==4.1.1
tiktoken==0.5.1